refactor: rename titleCase.ts to title-case.ts
Ref. metriport/metriport-internal#1040
